requirements.txt
aniso8601==1.1.0
click==6.2
Flask==0.10.1
Flask-Auth==0.85
Flask-JWT==0.3.2
Flask-PyMongo==0.4.0
Flask-RESTful==0.3.4
foreman==0.9.7
futures==3.0.3
gunicorn==19.4.1
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
ordereddict==1.1
PyJWT==1.4.0
pymongo==3.1.1
python-dateutil==2.4.2
python-dotenv==0.3.0
pytz==2015.7
requests==2.8.1
requests-futures==0.9.5
six==1.10.0
Werkzeug==0.11.2
wheel==0.24.0
